Minecraft-Pygame-main/requirements.txt
pyopengl==3.1.5
pyglet==1.5.14
numpy==1.19.3
pygame==2.0.1